Create skill component
