+++ src/sass/main.scss
@@ -4,7 +4,7 @@
 @import '../components/Menu/menu';
 @import '../components/Welcome/welcome';
 @import '../components/About/about';
-@import 'components/skills';
+@import '../components/Skills/skills';
 @import 'components/projects';
 @import 'base/global';
 @import '~bootstrap/scss/bootstrap';
